+++ components.json
@@ -5,12 +5,12 @@
   "tsx": true,
   "tailwind": {
     "config": "tailwind.config.ts",
-    "css": "app/globals.css",
+    "css": "src/app/globals.css",
     "baseColor": "slate",
     "cssVariables": true
   },
   "aliases": {
-    "components": "@/_components",
-    "utils": "@/_utils"
+    "components": "@/components",
+    "utils": "@/utils"
   }
 }
