+++ src/app/(auth)/components/header.tsx
