+++ src/app/(auth)/layout.tsx
@@ -1,5 +1,5 @@
-import { Footer } from "./_components/footer";
-import { Header } from "./_components/header";
+import { Footer } from "./components/footer";
+import { Header } from "./components/header";
 
 export default function AuthLayout({
   children,
