+++ src/app/(auth)/sign-up/page.tsx
@@ -1,6 +1,6 @@
-import { cn } from "@/_utils";
+import { cn } from "@/utils";
 import { Metadata } from "next";
-import { SignUpForm } from "../_components/sign-up-form";
+import { SignUpForm } from "../components/sign-up-form";
 
 export const metadata: Metadata = {
   title: "Welcome",
