+++ src/app/(marketing)/page.tsx
@@ -1,4 +1,4 @@
-import { Button } from "@/_components/ui/button";
+import { Button } from "@/components/ui/button";
 import Link from "next/link";
 
 export default function MarketingPage() {
