+++ src/components/ui/input.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { cn } from "@/_utils";
+import { cn } from "@/utils";
 
 export interface InputProps
   extends React.InputHTMLAttributes<HTMLInputElement> {}
