+++ src/components/ui/toaster.tsx
@@ -7,8 +7,8 @@ import {
   ToastProvider,
   ToastTitle,
   ToastViewport,
-} from "@/_components/ui/toast";
-import { useToast } from "@/_components/ui/use-toast";
+} from "@/components/ui/toast";
+import { useToast } from "@/components/ui/use-toast";
 
 export function Toaster() {
   const { toasts } = useToast();
